+++ HISTORY.md
@@ -1,3 +1,19 @@
+## 0.14.0 (2011-??-??)
+
+* Dot Notation: {{person.name}}
+* Implicit Iterators: {{# people}} {{.}} {{/ people}}
+* Automagic attr_readers for ivars you set in Sinatra routes.
+* Ability to customize escaping by overriding Mustache#escapeHTML
+
+## 0.13.0 (2011-02-23)
+
+* Mustache.render accepts a symbol to specify a template name.
+* Bugfix: Struct can be used as sections again
+
+## 0.12.1 (2011-02-22)
+
+* Bugfixes!
+
 ## 0.12.0 (2010-12-10)
 
 * Passing the full draft Mustache spec.